Restored backup-to-s3 script for PostgreSQL backup

+++ postgres/backup-to-s3.py
@@ -20,7 +20,10 @@ DBDATESTAMP = datetime.datetime.now().strftime('%Y%m%d_%H%M%S')
 #)
 
 # `postgis://` isn't recognized by `pg_dump`; replace it with `postgres://`
-DBURL = re.sub(r'^postgis://', 'postgres://', os.getenv('DATABASE_URL'))
+DBURL = re.sub(r'^postgis://', 'postgres://', os.getenv('DATABASE_URL')).replace(
+    os.getenv("POSTGRES_HOST"), "127.0.0.1"
+)
+
 DUMPFILE = 'postgres-{}-{}-{}.pg_dump'.format(
     os.environ.get('PG_MAJOR'),
     os.environ.get('PUBLIC_DOMAIN_NAME'),
@@ -28,26 +31,27 @@ DUMPFILE = 'postgres-{}-{}-{}.pg_dump'.format(
 )
 BACKUP_COMMAND = 'pg_dump --format=c --dbname="{}"'.format(DBURL)
 
-# If LifeCycle rules are enabled on bucket to move files to Glacier.
-# Keep in mind that file must stay in Glacier for (at least) 90 days.
-# Otherwise early deletion fee will be charged.
-# 7 days S3 Standard + 90 days in Glacier = 97 days
+yearly_retention = int(os.environ.get("AWS_BACKUP_YEARLY_RETENTION", 2))
+monthly_retention = int(os.environ.get("AWS_BACKUP_MONTHLY_RETENTION", 12))
+weekly_retention = int(os.environ.get("AWS_BACKUP_WEEKLY_RETENTION", 4))
+daily_retention = int(os.environ.get("AWS_BACKUP_DAILY_RETENTION", 30))
+
 DIRECTORIES = [
-    {'name': 'yearly', 'keeps': 2, 'days': 365},
-    {'name': 'monthly', 'keeps': 4, 'days': 30},
-    {'name': 'weekly', 'keeps': 14, 'days': 7},
-    {'name': 'daily', 'keeps': 97, 'days': 1},
+    {'name': 'postgres/yearly', 'keeps': yearly_retention, 'days': 365},
+    {'name': 'postgres/monthly', 'keeps': monthly_retention, 'days': 30},
+    {'name': 'postgres/weekly', 'keeps': weekly_retention, 'days': 7},
+    {'name': 'postgres/daily', 'keeps': daily_retention, 'days': 1},
 ]
 
 # Consider backups invalid whose (compressed) size is below this number of
 # bytes
-MINIMUM_SIZE = 100 * 1024 ** 2
+MINIMUM_SIZE = int(os.environ.get("AWS_POSTGRES_BACKUP_MINIMUM_SIZE", 100)) * 1024 ** 2
 
 # Data will be written directly to S3
 AWS_ACCESS_KEY_ID = os.environ.get('AWS_ACCESS_KEY_ID')
 AWS_SECRET_ACCESS_KEY = os.environ.get('AWS_SECRET_ACCESS_KEY')
 AWS_BUCKET = os.environ.get('BACKUP_AWS_STORAGE_BUCKET_NAME')
-CHUNK_SIZE = 250 * 1024 ** 2
+CHUNK_SIZE = int(os.environ.get("AWS_BACKUP_CHUNK_SIZE", 250)) * 1024 ** 2
 
 ###############################################################################
 
@@ -72,7 +76,7 @@ for directory in DIRECTORIES:
 
 # Perform the backup
 filename = ''.join((prefix, DUMPFILE))
-print 'Backing up to "{}"...'.format(filename)
+print('Backing up to "{}"...'.format(filename))
 upload = s3bucket.new_key(filename)
 chunks_done = 0
 with smart_open.smart_open(upload, 'wb') as s3backup:
@@ -81,27 +85,29 @@ with smart_open.smart_open(upload, 'wb') as s3backup:
     while True:
         chunk = process.stdout.read(CHUNK_SIZE)
         if not len(chunk):
-            print 'Finished! Wrote {} chunks; {}'.format(
+            print('Finished! Wrote {} chunks; {}'.format(
                 chunks_done,
                 humanize.naturalsize(chunks_done * CHUNK_SIZE)
-            )
+            ))
             break
         s3backup.write(chunk)
         chunks_done += 1
         if not '--hush' in sys.argv:
-            print 'Wrote {} chunks; {}'.format(
+            print('Wrote {} chunks; {}'.format(
                 chunks_done,
                 humanize.naturalsize(chunks_done * CHUNK_SIZE)
-            )
+            ))
 
-# Remove old backups beyond desired retention
-for directory in DIRECTORIES:
-    prefix = directory['name'] + '/'
-    keeps = directory['keeps']
-    s3keys = s3bucket.list(prefix=prefix)
-    large_enough_backups = filter(lambda x: x.size >= MINIMUM_SIZE, s3keys)
-    large_enough_backups = sorted(large_enough_backups, key=lambda x: x.last_modified, reverse=True)
+aws_lifecycle = os.environ.get("AWS_BACKUP_BUCKET_DELETION_RULE_ENABLED", "False") == "True"
+if not aws_lifecycle:
+    # Remove old backups beyond desired retention
+    for directory in DIRECTORIES:
+        prefix = directory['name'] + '/'
+        keeps = directory['keeps']
+        s3keys = s3bucket.list(prefix=prefix)
+        large_enough_backups = filter(lambda x: x.size >= MINIMUM_SIZE, s3keys)
+        large_enough_backups = sorted(large_enough_backups, key=lambda x: x.last_modified, reverse=True)
 
-    for l in large_enough_backups[keeps:]:
-        print 'Deleting old backup "{}"...'.format(l.name)
-        l.delete()
+        for l in large_enough_backups[keeps:]:
+            print('Deleting old backup "{}"...'.format(l.name))
+            l.delete()
